добавлен максимальный размер картинки

+++ Bazzar/picture-service/src/main/java/ru/bazzar/picture/service/PictureService.java
@@ -59,8 +59,11 @@ public class PictureService {
 
 
     public void deleteById(Long id){
-        pictureRepository.deleteById(id);
+        if(id != null && id > 1){
+            pictureRepository.deleteById(id);
+        }
     }
+
 
 
 }
